pyproject.toml
[tool.poetry]
name = "geo-points-api"
version = "0.1.0"
description = "Backend-приложение на Django для работы с географическими точками"
authors = ["Evgeniy Slizh"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
djangorestframework = "^3.14"
django-cors-headers = "^4.3"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
pytest-cov = "^4.1"
black = "^23.11"
isort = "^5.12"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

